requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.0
filelock==3.14.0
Flask==3.0.0
Flask-WTF==1.2.1
frozenlist==1.4.1
fsspec==2024.6.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.23.3
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
litellm==1.40.6
MarkupSafe==2.1.3
multidict==6.0.5
numpy==1.24.4
openai==1.33.0
packaging==23.2
pandas==1.5.3
pydantic==2.5.3
pydantic_core==2.14.6
pymongo==4.7.3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
six==1.16.0
sniffio==1.3.0
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.2.1
Werkzeug==3.0.1
WTForms==3.1.1
yarl==1.9.4
zipp==3.17.0
